requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
anthropic
python-multipart
aiofiles
Pillow
PyJWT
httpx
redis
pywebpush
firebase-admin
